requirement.txt
asgiref==3.4.1
azure-core==1.24.2
azure-identity==1.10.0
cachetools==4.2.4
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.0.4
contextvars==2.4
cryptography==37.0.4
Django==3.2.6
django-common-helpers==0.9.2
django-cors-headers==3.7.0
django-cron==0.5.1
djangorestframework==3.12.4
google-api-core==2.8.2
google-auth==2.11.0
googleapis-common-protos==1.56.3
idna==3.2
immutables==0.18
msal==1.18.0
msal-extensions==1.0.0
mysqlclient==2.0.3
numpy==1.19.5
opencensus==0.11.0
opencensus-context==0.1.3
opencensus-ext-azure==1.1.7
pandas==1.1.5
portalocker==2.5.1
protobuf==3.19.4
psutil==5.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
rsa==4.9
six==1.16.0
sqlparse==0.4.1
typing_extensions==4.1.1
urllib3==1.26.6
